requirements.txt
sqlalchemy
fastapi
uvicorn
pydantic-settings
asyncpg
pytest
httpx
ortools
humanfriendly
pandas
psycopg2-binary
SQLAlchemy-Utc